[FEAT] Decay Changes (#14)
* first attempt at rot ability.

* revamped rot decay

* switched from ray casting to a full fill

* rot is really cool now

* increased max blocks per tick so that it removes striping

* rot now can be sped uo

* better control over quirk factor multipliers

* adding a wip environmental decay

* environmental decay is awesome now

* reduced particle count

* speed changes

* renamed decay types and fixed speed calc

* added vein miner ability

* added fissure mode

* added decay mob effect to library

* fixed crash because of body_condition

When the player dies, the body capability becomes unavailable, but the body conditions for the power still fire. So added a graceful fix

* fixed similar crash to body_status

just like the body status helper, this helper would crash the game if there is a power using the stamina condition, and the player dies.

* properly destroys item in hand

* apply decay effect to those in range

* lowered particles and range for rot

* added logic to restrict movement with tag

Not sure if I am ever gonna use this, but adding it for now

* made connected decay drop items occasionally

* added new grab ability

Adds in a hold and a telekinesis grab. Needs some work because the hold doesn't stop the user's movement like it should.

* added stamina cost to env decay

* attempt at mixing in to Ability.class

Failed attempt so far. Pushing to share progress

* started AbilityInstance mixin to attempt do call the tick command

* work in progress

* bad polymorphism attempt -_-

* finally, a working universal stamina system!

* modified tick order and added tick() and lastTick()

still an issue where it drains all your stamina when you log in

* changed stamina to reset usage total correctly

* modified stamina properties

Added a new first tick only cost
Modified to stamina_interval and stamina_interval_cost

* added powers enabled condition

* removed unnecessary prints

* Grab now freezes the user as well

* added a stun effect

* changed grab to use new stun when ability is released

* scaled the potion effect with quirk factor

* stamina changes.

* let user know if their decay fails due to intensity

* removed duplicate command and exposed quirk helper

* deleting unused command for review

---------

Co-authored-by: Bandit <174976884+BanditHelps@users.noreply.github.com>